# lib/api-zod/src/generated/types/discoveryAsset.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Solana discovery, Houdini private swaps, and separate OKX public swaps
 * OpenAPI spec version: 0.1.0
 */

export interface DiscoveryAsset {
  assetId: string;
  name: string;
  symbol: string;
  category?: string;
  mint?: string;
  icon?: string;
  price?: number;
  change24h?: number;
  volume24h?: number;
  marketCap?: number;
  liquidity?: number;
  trustTier?: string;
}