# lib/api-zod/src/generated/types/discoveryDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Solana discovery, Houdini private swaps, and separate OKX public swaps
 * OpenAPI spec version: 0.1.0
 */
import type { DiscoveryAsset } from './discoveryAsset';
import type { DiscoveryDetailMetadataStatus } from './discoveryDetailMetadataStatus';

export interface DiscoveryDetail {
  asset: DiscoveryAsset;
  description?: string;
  supply?: number;
  decimals?: number;
  mintAuthority?: string;
  freezeAuthority?: string;
  onChainName?: string;
  onChainSymbol?: string;
  metadataStatus?: DiscoveryDetailMetadataStatus;
}