# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Solana discovery, Houdini private swaps, and separate OKX public swaps
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}