# lib/api-zod/src/generated/types/okxQuote.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Solana discovery, Houdini private swaps, and separate OKX public swaps
 * OpenAPI spec version: 0.1.0
 */

export interface OkxQuote {
  fromMint: string;
  toMint: string;
  amountIn: string;
  amountOut: string;
  fromSymbol: string;
  toSymbol: string;
  priceImpactPercent: string;
  tradeFeeUsd?: string;
}